+++ gea/libs/gea_math/src/gea/mth_vector/vector3.inl
@@ -127,7 +127,7 @@ inline const vector3 vector3::cross_product(const vector3 &o) const {
 inline const float vector3::angle(const vector3 &o) const {
     // theta = arccos( a dot b / |a||b| )
     const float len = length() * o.length();
-    return (len != 0.0f) ? acos(dot_product(o) / len) : 0.0f;
+    return (len != 0.0f) ? acosf(dot_product(o) / len) : 0.0f;
 }
 
 // attributes
